fixed access point bug and refined the access point service uninstaller

+++ bin/access-point/stop-access-point-service.sh
@@ -10,12 +10,19 @@ systemctl stop hostapd
 echo 'stoped hostapd'
 
 # restore the dhcpcd config file
-mv $DHCPCD_CONF.save $DHCPCD_CONF
-echo "$DHCPCD_CONF file restored"
+if [ -f $DHCPCD_CONF.save ]; then
+  mv $DHCPCD_CONF.save $DHCPCD_CONF
+  echo "$DHCPCD_CONF file restored"
+else
+  echo "warning: $DHCPCD_CONF.save file could not be found"
+  echo "  manually restore $DHCPCD_CONF"
+fi
 
 # restore the network interfaces config file
-mv $INTERFACES.save $INTERFACES
-echo "$INTERFACES file restored"
-
-echo "Please reboot for changes to take effect"
-
+if [ -f $INTERFACES.save ]; then
+  mv $INTERFACES.save $INTERFACES
+  echo "$INTERFACES file restored"
+else
+  echo "warning: $INTERFACES.save file could not be found"
+  echo "  manually restore $INTERFACES"
+fi
